Calling tag api for creating tag assignment with issues

+++ src/main/java/com/dev/issuebook/entity/IssueEntity.java
@@ -21,7 +21,7 @@ public class IssueEntity implements Serializable {
     private Long id;
 
 	@Column(name = "user_id", nullable = false)
-	private Integer userId;
+	private Long userId;
 
 	@Column(name = "issue_desc", nullable = false, columnDefinition = "TEXT")
 	private String issueDesc;
@@ -47,7 +47,7 @@ public class IssueEntity implements Serializable {
 	public IssueEntity() {
 	}
 
-	public IssueEntity(Long id, Integer userId, String issueDesc, String issueType, String rootCause,
+	public IssueEntity(Long id, Long userId, String issueDesc, String issueType, String rootCause,
 			String resolution, Boolean resolved, String refs, LocalDateTime createdAt) {
 		this.id = id;
 		this.userId = userId;
@@ -68,11 +68,11 @@ public class IssueEntity implements Serializable {
 		this.id = id;
 	}
 
-	public Integer getUserId() {
+	public Long getUserId() {
 		return userId;
 	}
 
-	public void setUserId(Integer userId) {
+	public void setUserId(Long userId) {
 		this.userId = userId;
 	}
